Add url_value_preprocessor and url_defaults to dashboard blueprint
Since things like username and budget_id will be used heavily throughout the blueprint,
this essentially scrapes them out of the url the first time they are entered and saves them off into
g which can then be used for providing url defaults when calling url_for.

Still have to be careful and provide defaults for things like budget_id when not entered - such as the last
visited budget or the first budget a user created as a fallback.