Convert authenticate function into before_request function in dashboard blueprint
url_defaults and url_value_preprocessor will ensure the username and budget_id is preserved in
the global state. This then allows the before_request to occur before every request in dashboard
in order to check that the current user is only trying to look at their user data and only at
budgets and accounts owned by themselves.